cmd: reorganize network file structure
