+++ starport/interface/cli/starport/cmd/network_chain.go
@@ -1,0 +1,12 @@
+package starportcmd
+
+import "github.com/spf13/cobra"
+
+func NewNetworkChain() *cobra.Command {
+	c := &cobra.Command{
+		Use:  "chain",
+		Args: cobra.ExactArgs(1),
+	}
+	c.AddCommand(NewNetworkChainCreate())
+	return c
+}
